ml_backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
scikit-learn
slowapi
python-dotenv
pytest
requests
python-multipart
httpx
